Configurable status LED
